Remove django-compressor
