+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "WebODM",
-  "version": "2.4.0",
+  "version": "2.4.1",
   "description": "User-friendly, extendable application and API for processing aerial imagery.",
   "main": "index.js",
   "scripts": {
