+++ requirements.txt
@@ -8,7 +8,6 @@ Django==2.2.27
 django-appconf==1.0.2
 django-codemirror2==0.2
 django-colorfield==0.1.15
-django-compressor==2.2
 django-cors-headers==3.0.2
 django-filter==2.4.0
 django-guardian==1.4.9
